Move the unused CharFilter to TestBugInSomething
